remove patch_linear and integrate matmul allreduce in linear.py together
