+++ vllm_ascend/patch/worker/patch_common/__init__.py
@@ -16,7 +16,6 @@
 #
 
 import vllm_ascend.patch.worker.patch_common.patch_distributed  # noqa
-import vllm_ascend.patch.worker.patch_common.patch_linear  # noqa
 import vllm_ascend.patch.worker.patch_common.patch_logits  # noqa
 import vllm_ascend.patch.worker.patch_common.patch_lora_embedding  # noqa
 import vllm_ascend.patch.worker.patch_common.patch_minicpm  # noqa
